requirements.txt
selenium
webdriver_manager
python-dotenv
beautifulsoup4